* collision now uses size_t instead of unsigned int for hashing objects. this solves the problems where sometimes hashes would collide under 64-bit platforms and cause crashes, and/or other weird stuff to happen. * please test and give feedback, especially if you are running under x64 and using collision (a full rebuild of the collide library is required)
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4074 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ extras/collision/collision_dll/g3dlite/G3D/Triangle.h
@@ -96,7 +96,7 @@ public:
         return true;
     }
 
-    inline unsigned int hashCode() const {
+    inline size_t hashCode() const {
         return
             _vertex[0].hashCode() +
             (_vertex[1].hashCode() >> 2) +
@@ -109,7 +109,7 @@ public:
 
 } // namespace
 
-inline unsigned int hashCode(const G3D::Triangle& t) {
+inline size_t hashCode(const G3D::Triangle& t) {
 	return t.hashCode();
 }
